Add usb-reset library.
This allows to issue a software reset to a nonresponsive USB device.
It requires write access on the USB bus device node so add a udev rule
for the current meter USB TTY device.
Put the reset code into its own crate to allow using "forbid_unsafe" in the
main code.