Improve appliance controller behaviour.
It behaved incorrectly when when multiple appliances were chained.

As solution, pass around signed "available" powers in the chain. Then increase
or decrease the appliance by the passed amount. The result is some kind
of "I" controller which uses the measured appliance power instead of an
integration register.

The first appliance in the chain is adjusted first in both directions,
the last appliance is adjusted last.